feat: introduce AuthTypes and refactor UIWindow extension for improved authentication handling
