+++ Sources/FronteggSwift/services/StepUpAuthenticator.swift
@@ -56,12 +56,11 @@ class StepUpAuthenticator {
             }
         }
 
-        let (authorizeUrl, codeVerifier) = AuthorizeUrlGenerator.shared.generate(
+        let (authorizeUrl, _) = AuthorizeUrlGenerator.shared.generate(
             stepUp: true,
             maxAge: maxAge
         )
         
-        CredentialManager.saveCodeVerifier(codeVerifier)
         DispatchQueue.main.async {
             FronteggAuth.shared.setIsStepUpAuthorization(true)
             FronteggAuth.shared.setIsLoading(false)
